refactor : create writer/ and reader/
