+++ lib/object.js
@@ -8,8 +8,8 @@ var compressions = require('./compressions');
 var CompressedObject = require('./compressedObject');
 var nodeBuffer = require('./nodeBuffer');
 var utf8 = require('./utf8');
-var StringWriter = require('./stringWriter');
-var Uint8ArrayWriter = require('./uint8ArrayWriter');
+var StringWriter = require('./writer/stringWriter');
+var Uint8ArrayWriter = require('./writer/uint8ArrayWriter');
 var ZipObject = require('./zipObject');
