+++ lib/reader/dataReader.js
@@ -1,5 +1,5 @@
 'use strict';
-var utils = require('./utils');
+var utils = require('../utils');
 
 function DataReader(data) {
     this.data = null; // type : see implementation
