+++ lib/reader/nodeBufferReader.js
