+++ lib/reader/readerFor.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var utils = require('./utils');
-var support = require('./support');
+var utils = require('../utils');
+var support = require('../support');
 var StringReader = require('./stringReader');
 var NodeBufferReader = require('./nodeBufferReader');
 var Uint8ArrayReader = require('./uint8ArrayReader');
