+++ lib/reader/stringReader.js
@@ -1,6 +1,6 @@
 'use strict';
 var DataReader = require('./dataReader');
-var utils = require('./utils');
+var utils = require('../utils');
 
 function StringReader(data, optimizedBinaryString) {
     this.data = data;
