+++ lib/reader/uint8ArrayReader.js
