+++ lib/writer/stringWriter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('./utils');
+var utils = require('../utils');
 
 /**
  * An object to write any content to a string.
