+++ lib/writer/uint8ArrayWriter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('./utils');
+var utils = require('../utils');
 
 /**
  * An object to write any content to an Uint8Array.
