+++ lib/zipEntry.js
@@ -1,5 +1,5 @@
 'use strict';
-var readerFor = require('./reader');
+var readerFor = require('./reader/readerFor');
 var utils = require('./utils');
 var CompressedObject = require('./compressedObject');
 var jszipProto = require('./object');
